+++ core-services/zuul/src/main/java/org/egov/contract/Role.java
@@ -12,7 +12,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @EqualsAndHashCode(of = {"code", "tenantId"})
-public class Role {
+public class Role implements java.io.Serializable {
+
+    private static final long serialVersionUID = 8967926454506815772L;
+
     @JsonProperty("id")
     private Long id;